requirements.txt
Django==5.2.4
django-colorfield==0.11.0
django-admin-interface==0.26.0
djangorestframework==3.15.0
django-cors-headers==4.3.1
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9 